% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_Numeric_Categorical.R
\name{boxplot_numeric_category}
\alias{boxplot_numeric_category}
\title{Multiple Boxplot for a Numeric Variable across Categories}
\usage{
boxplot_numeric_category(dataframe, x, y)
}
\value{
A single panel multiple boxplot showing the distribution of the variable y in each category contained in the variable x.
}
\description{
This function creates a multiple boxplot that shows the distribution of a numeric variable in each different category contained in a categorical variable. This function will remove any NA values in the data prior to creation of the boxplot.
}
\examples{
using the "CO2" dataset from the r "datasets" package, create a boxplot that shows the distribution of CO2 concentration of grass plants in each treatment type (chilled or nonchilled)
boxplot_numeric_category(dataframe = CO2, x = Treatment, y = conc)
#using the "starwars" dataset from the dplyr package,create a boxplot that shows the distribution of the height of starwars characters based on the sex of the characters
starwars \%>\% boxplot_numeric_category(sex, height)
#using the "storms" dataset from the dplyr package, create a boxplot that shows the distribution of the wind speeds observed for each of the storms Caroline and Doris
boxplot_numeric_category((storms \%>\% filter(name == c("Amy", "Doris"))), name, wind)
}
