% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{model_sign}
\alias{model_sign}
\title{pvalue}
\usage{
model_sign(x)
}
\arguments{
\item{x}{An object of class 'lm', your linear model}

\item{ifelse()}{Logical statement which tests a statement about the object in the brackets}

\item{p}{Returns the p-value for the slope coefficient}

\item{q}{Interprets your p-value at 95\% confidence level}

\item{r}{Returns the adjusted R-squared value for your model}

\item{s}{Interprets the adjusted R-squared value for your model}

\item{summary()}{Function that generates the summary table with results from your fitted model}

\item{coefficients[]}{Extracts the p-value for the slope coefficient from the summary table}
}
\value{
A vector containing p,q,r and s
}
\description{
Interprets the p-value and adjusted r-squared for a regression model
}
\examples{
pvalue(lm(mtcars$hp ~ mtcars$wt))
}
