% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicalParam.R
\name{poten_logi_param}
\alias{poten_logi_param}
\title{Logical Variables Identification}
\usage{
poten_logi_param(dataset, logical_factors = 2)
}
\arguments{
\item{dataset}{the data frame/tibble to be checked for logical variables}

\item{logical_factors}{Number of factors to be considered for logical variable.
Default value is 2.}
}
\value{
character vector of column names from the dataset
}
\description{
Description: create a character vector of column names of the given
data frame which can be identified as logical in nature
}
