% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Generate p-value}
\usage{
pvalue(x)
}
\arguments{
\item{x}{An object of class 'lm', your linear model}

\item{if()}{Logical statement which tests a statement about the object in the brackets}

\item{p}{Returns the p-value for the slope coefficient}

\item{summary()}{Function that generates the summary table with results from your fitted model}

\item{coefficients[]}{Extracts the p-value for the slope coefficient from the summary table}
}
\value{
P-value for slope coefficient
}
\description{
Returns the p-value for a simple linear regression with 1 independent variable
}
\examples{
pvalue(reg1)
pvalue(lm(mtcars$hp ~ mtcars$wt))
}
