% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPackageMC1-package.R
\docType{package}
\name{rPackageMC1-package}
\alias{rPackageMC1}
\alias{rPackageMC1-package}
\title{rPackageMC1: Create a list of potential logical variables}
\description{
This package utilizes a function to identify the logical variables. By default, logical variable has two factors, and this package helps identify the variables in a dataframe which can be potentially a logical variable.
}
\author{
\strong{Maintainer}: Mrinmoy Chakraborty \email{mrinmoy.chakraborty@ubc.ca} (\href{https://orcid.org/0000-0002-4947-9265}{ORCID})

}
\keyword{internal}
